docs: Docs as stories (WIP)
