+++ packages/examples-tab/package.json
@@ -3,30 +3,10 @@
   "version": "5.3.0-alpha.2",
   "private": true,
   "dependencies": {
-    "@emotion/core": "^10.0.9",
-    "react": "^16.8.4"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.4.0",
-    "@babel/plugin-transform-react-jsx": "^7.3.0",
-    "@babel/preset-env": "^7.4.2",
-    "@storybook-addon-designs/assets": "^5.3.0-alpha.0",
-    "@storybook/addons": "^5.3.8",
-    "@storybook/react": "^5.3.8",
-    "@storybook/theming": "^5.3.8",
-    "babel-loader": "^8.0.5",
-    "storybook-addon-designs": "^5.3.0-alpha.2"
+    "examples": "5.3.0-alpha.2"
   },
   "scripts": {
     "storybook": "start-storybook -p 6007 -s ../assets",
     "build-storybook": "build-storybook -s ../assets"
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env"
-    ],
-    "plugins": [
-      "@babel/plugin-transform-react-jsx"
-    ]
   }
 }
