+++ packages/examples/.storybook/main.js
@@ -1,4 +1,4 @@
 module.exports = {
-  stories: ['../stories/**/*.stories.{js,jsx}'],
-  addons: ['storybook-addon-designs']
+  stories: ['../stories/**/*.stories.mdx', '../stories/**/*.stories.{js,jsx}'],
+  addons: ['@storybook/addon-docs', 'storybook-addon-designs']
 }
