+++ packages/examples/.storybook/preview.js
@@ -1,6 +1,8 @@
 import { addParameters } from '@storybook/react'
 
 addParameters({
+  viewMode: 'story',
+  docs: { page: null },
   options: {
     showRoots: true
   }
