+++ packages/examples/package.json
@@ -11,11 +11,12 @@
     "@babel/plugin-transform-react-jsx": "^7.3.0",
     "@babel/preset-env": "^7.4.2",
     "@storybook-addon-designs/assets": "^5.3.0-alpha.0",
-    "@storybook/addon-actions": "^5.3.8",
-    "@storybook/addon-links": "^5.3.8",
-    "@storybook/addons": "^5.3.8",
-    "@storybook/react": "^5.3.8",
-    "@storybook/theming": "^5.3.8",
+    "@storybook/addon-actions": "5.3.18",
+    "@storybook/addon-docs": "5.3.18",
+    "@storybook/addon-links": "5.3.18",
+    "@storybook/addons": "5.3.18",
+    "@storybook/react": "5.3.18",
+    "@storybook/theming": "5.3.18",
     "babel-loader": "^8.0.5",
     "storybook-addon-designs": "^5.3.0-alpha.2"
   },
